requirements.txt
streamlit==1.37.0
pillow==10.3.0
numpy==1.26.4
tflite-runtime==2.14.0
